bump offsets to 64bits

+++ storage/backends/plakard/client.go
@@ -349,7 +349,7 @@ func (repository *Repository) GetPackfile(checksum objects.Checksum) (io.Reader,
 	return bytes.NewBuffer(data), nil
 }
 
-func (repository *Repository) GetPackfileBlob(checksum objects.Checksum, offset uint32, length uint32) (io.Reader, error) {
+func (repository *Repository) GetPackfileBlob(checksum objects.Checksum, offset uint64, length uint32) (io.Reader, error) {
 	result, err := repository.sendRequest("ReqGetPackfileBlob", network.ReqGetPackfileBlob{
 		Checksum: checksum,
 		Offset:   offset,
